Math: Added pow and sqrt
Added power function, square root function, edited abs to fabs

+++ src/CLib/math/fabs.c
@@ -1,6 +1,6 @@
 #include "math.h"
 
-int abs(int a)
+float abs(float a)
 {
     return a < 0 ? -a : a;
 }
